Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/referrer-policy/same-origin/attr-referrer/same-origin/http-http/img-tag/swap-origin-redirect/same-origin-insecure.http-expected.txt



FAIL The referrer URL is omitted when a
                                 document served over http requires an http
                                 sub-resource via img-tag using the attr-referrer
                                 delivery method with swap-origin-redirect and when
                                 the target request is same-origin. assert_equals: Reported Referrer URL is 'omitted'. expected (undefined) undefined but got (string) "http://localhost:8800/referrer-policy/same-origin/attr-referrer/same-origin/http-http/img-tag/swap-origin-redirect/same-origin-insecure.http.html"
FAIL The referrer URL is omitted when a
                                 document served over http requires an http
                                 sub-resource via img-tag using the attr-referrer
                                 delivery method with swap-origin-redirect and when
                                 the target request is same-origin. (srcdoc iframe inherits parent) assert_equals: Reported Referrer URL is 'omitted'. expected (undefined) undefined but got (string) "http://localhost:8800/referrer-policy/same-origin/attr-referrer/same-origin/http-http/img-tag/swap-origin-redirect/same-origin-insecure.http.html"
PASS The referrer URL is omitted when a
                                 document served over http requires an http
                                 sub-resource via img-tag using the attr-referrer
                                 delivery method with swap-origin-redirect and when
                                 the target request is same-origin. (overridden by srcdoc iframe)
